Создание сервиса MyService, перенос запроса контактов в сервис.
